Split out header file into it's own
